# Improve performance of JSON HTML entity escaping