@byroot 2023-07-05
So if I read the benchmark correctly, the AS::JSON overhead over stdlib JSON.generate is still about 4x?

Any hints as to where that time is spent? How much is escaping vs calling `as_json`?