@jhawthorn 2023-07-05
> So if I read the benchmark correctly, the AS::JSON overhead over stdlib JSON.generate is still about 4x?
> 
> Any hints as to where that time is spent? How much is escaping vs calling `as_json`?

At least based on this benchmark I'd say 1/2 `jsonify`, 1/4 escaping, and the rest vanilla JSON generation.

```
$ be stackprof /tmp/jhawthorn/stackprof20230705-56125-68l0vd.dump
==================================
  Mode: wall(1000)
  Samples: 12834 (11.94% miss rate)
  GC: 547 (4.26%)
==================================
     TOTAL    (pct)     SAMPLES    (pct)     FRAME
      6487  (50.5%)        4828  (37.6%)     ActiveSupport::JSON::Encoding::JSONGemEncoder#jsonify
      3336  (26.0%)        3336  (26.0%)     String#gsub!
      2090  (16.3%)        2090  (16.3%)     JSON::Ext::Generator::State#generate
      6491  (50.6%)         741   (5.8%)     Hash#each
       520   (4.1%)         520   (4.1%)     Module#===
       282   (2.2%)         282   (2.2%)     (sweeping)
       271   (2.1%)         271   (2.1%)     (marking)
       347   (2.7%)         253   (2.0%)     Kernel#require
       148   (1.2%)         148   (1.2%)     NilClass#===
       111   (0.9%)         111   (0.9%)     TrueClass#===
      6491  (50.6%)          73   (0.6%)     Array#map
        66   (0.5%)          66   (0.5%)     FalseClass#===
```

To improve jsonify from here I think we could try to avoid recreating objects we don't need to (hashes and arrays which are already deeply JSON-ready), but that's a bit tricky. We also have some unavoidable cost in iterating the tree, which YJIT may be able to help with. I previously targeted this in https://github.com/ruby/ruby/pull/6674 but something similar but I suspect not quite as fast was merged (could be worth revisiting).