@byroot 2023-07-05
Right, so the bulk is having to walk down the tree in Ruby.

I wonder if the "yield based" interface I added to RapidJSON helps there or not. We could try to contribute a similar API to flori/json. Would probably take a while to be accepted (if at all) but could lead to nice gains longer term.