@jhawthorn 2023-07-05
> I wonder if the "yield based" interface I added to RapidJSON helps there or not. We could try to contribute a similar API to flori/json. Would probably take a while to be accepted (if at all) but could lead to nice gains longer term.

I mean in this case it completely eliminates the iteration, so it's 100% faster 😅. I do think we should try to get a similar API more widely adopted if possible.