@byroot 2023-07-05
> I mean in this case it completely eliminates the iteration, so it's 100% faster 😅

Lol. Sure, but it also requires more calls from C into Ruby etc, so it's not so evident. 

But yeah, we should try to get this in flori/json. One of my goals with RapidJSON was to use it as a test bed for these stuff.