@nobu 2023-08-18
Maybe `tr!` isn't more faster?